+++ src/interface/IVaultManager.sol
@@ -5,9 +5,9 @@ import "./ILedgerComponent.sol";
 
 interface IVaultManager is ILedgerComponent {
     // get balance
-    function getBalance(uint256 _chainId, bytes32 _symbol) external view returns (uint256);
+    function getBalance(uint256 _chainId, bytes32 _symbol) external view returns (uint128);
     // add balance
-    function addBalance(uint256 _chainId, bytes32 _symbol, uint256 _deltaBalance) external;
+    function addBalance(uint256 _chainId, bytes32 _symbol, uint128 _deltaBalance) external;
     // sub balance
-    function subBalance(uint256 _chainId, bytes32 _symbol, uint256 _deltaBalance) external;
+    function subBalance(uint256 _chainId, bytes32 _symbol, uint128 _deltaBalance) external;
 }
